ai/requirements.txt
# =============================================================================
# LiveNote AI Gateway - Consolidated Requirements for Docker
# =============================================================================

# FastAPI & Server
fastapi==0.115.0
uvicorn[standard]==0.30.6

# HTTP Clients
httpx[http2]==0.27.2
requests>=2.31.0
aiohttp>=3.9.0
h2>=4.1.0

# OpenAI
openai==1.54.0

# Vector Database
chromadb>=1.0.20

# LangChain (RAG 모듈용)
langchain-core>=0.3,<0.4
langchain-community>=0.3,<0.4
langchain-openai>=0.2,<0.3
langchain-text-splitters>=0.3,<0.4
langchain-chroma>=0.2,<0.3

# Data Validation & Utilities
pydantic==2.9.0
python-dotenv==1.0.1
pandas==2.2.2

# PDF Processing
pypdf==4.2.0
reportlab==4.2.0

# File Upload
python-multipart==0.0.9

# Media Providers
youtube-transcript-api==0.6.1
rapidfuzz==3.6.1

# Date/Time Utilities
python-dateutil>=2.8.0

# Optional (for development/testing)
# pytest>=8.0,<9
# pytest-asyncio>=0.21.0
# pytest-cov>=4.0,<5
# streamlit==1.36.0
trio==0.32.0
